news_board/requirements.txt
asgiref==3.3.0
Django==3.0
pytz==2020.4
sqlparse==0.4.1
djangorestframework==3.12.4
psycopg2-binary==2.8.6
django-cors-headers==3.10.0